refactor : `View` 확장함수 `SnackBar` 관련 코드 `ViewExtensions` 패키지로 이동
